# docs/openapi.yaml
openapi: 3.0.3
info:
  title: KeenOn Card Generate API
  description: |
    API documentation for the KeenOn Card Generate service.
    This service provides endpoints for user authentication, card generation, and folder management.
  version: 1.0.0
  contact:
    name: KeenOn Development Team
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.keenon-card-generate.example.com
    description: Production server
...
paths:
...
  /api/cards:
    get:
      tags:
        - Cards
      summary: Get all cards
      description: Get all cards for the authenticated user with pagination
      operationId: getCards
      security:
        - cookieAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: Page number
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
          description: Number of items per page
        - name: folderId
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by folder ID
      responses:
        '200':
          description: Cards retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  results:
                    type: integer
                    example: 10
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Card'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - Cards
      summary: Create a card
      description: Create a new card for the authenticated user
      operationId: createCard
      security:
        - cookieAuth: []
        - csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - word
                - translation
                - image
              properties:
                word:
                  type: string
                  example: 你好
                translation:
                  type: string
                  example: Hello
                image:
                  type: string
                  format: binary
                  description: Base64 encoded image
                folderId:
                  type: string
                  format: uuid
                  description: Folder ID to place the card in
      responses:
        '201':
          description: Card created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/Card'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/folders:
    get:
      tags:
        - Folders
      summary: Get all folders
      description: Get all folders for the authenticated user
      operationId: getFolders
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Folders retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Folder'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - Folders
      summary: Create a folder
      description: Create a new folder for the authenticated user
      operationId: createFolder
      security:
        - cookieAuth: []
        - csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  example: Chinese Basics
                description:
                  type: string
                  example: Basic Chinese vocabulary
      responses:
        '201':
          description: Folder created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/Folder'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        name:
          type: string
          example: John Doe
        email:
          type: string
          format: email
          example: john.doe@example.com
        verified:
          type: boolean
          example: true
        role:
          type: string
          enum: [user, admin]
          example: user
        createdAt:
          type: string
          format: date-time
          example: 2023-01-01T00:00:00Z
        updatedAt:
          type: string
          format: date-time
          example: 2023-01-01T00:00:00Z
    Card:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        word:
          type: string
          example: 你好
        translation:
          type: string
          example: Hello
        image:
          type: string
          format: binary
          description: Base64 encoded image
        characterBreakdown:
          type: array
          items:
            type: string
          example: ["你 - you", "好 - good"]
        exampleSentences:
          type: array
          items:
            type: string
          example: ["你好吗? - How are you?"]
        user:
          $ref: '#/components/schemas/User'
        folder:
          $ref: '#/components/schemas/Folder'
        createdAt:
          type: string
          format: date-time
          example: 2023-01-01T00:00:00Z
        updatedAt:
          type: string
          format: date-time
          example: 2023-01-01T00:00:00Z
    Folder:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        name:
          type: string
          example: Chinese Basics
        description:
          type: string
          example: Basic Chinese vocabulary
        user:
          $ref: '#/components/schemas/User'
        cards:
          type: array
          items:
            $ref: '#/components/schemas/Card'
        createdAt:
          type: string
          format: date-time
          example: 2023-01-01T00:00:00Z
        updatedAt:
          type: string
          format: date-time
          example: 2023-01-01T00:00:00Z
    Error:
      type: object
      properties:
        status:
          type: string
          example: error
        message:
          type: string
          example: An error occurred
        code:
          type: string
          example: INTERNAL_SERVER_ERROR
        stack:
          type: string
          description: Only included in development environment
          example: Error at line 42...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
    csrfToken:
      type: apiKey
      in: header
      name: csrf-token